- move dtos to a model.transport package
